pyproject.toml
[project]
name = "jr-west-delay-consecutive-days"
version = "0.1.0"
description = "JR西日本の各路線ごとの遅延連続日数カウンター"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
    "tweepy>=4.14.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 99
target-version = "py311"
